requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2
requests==2.31.0
docker==6.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
tenacity==8.2.3